requirements.txt
pyserial==3.5
requests
numpy
schedule
pyttsx3